Import lua syntax instead of copying it
The previous version copied an (older) version of the lua syntax file and then made modifications to it.
It was difficult to tell which parts were specific to pico8.
Now we will start off by importing the lua syntax file, and layer pico8 changes on top of it.
This also allows benefiting from any improvements in the lua syntax file over time.

It _does_ create a dependency on the lua file, so users will need to make sure they have
lua syntax installed before installing this plugin.